StockAnalyzer/requirements.txt
streamlit
yfinance
matplotlib
pandas
requests
gtts
python-dotenv
